Move the debug overlay rendering out from EmuScreen into DebugOverlay.cpp
