+++ UI/UI.vcxproj
@@ -40,6 +40,7 @@
     <ClCompile Include="ControlMappingScreen.cpp" />
     <ClCompile Include="CustomButtonMappingScreen.cpp" />
     <ClCompile Include="CwCheatScreen.cpp" />
+    <ClCompile Include="DebugOverlay.cpp" />
     <ClCompile Include="DevScreens.cpp" />
     <ClCompile Include="DiscordIntegration.cpp" />
     <ClCompile Include="DisplayLayoutScreen.cpp" />
@@ -76,6 +77,7 @@
     <ClInclude Include="ChatScreen.h" />
     <ClInclude Include="ControlMappingScreen.h" />
     <ClInclude Include="CustomButtonMappingScreen.h" />
+    <ClInclude Include="DebugOverlay.h" />
     <ClInclude Include="DevScreens.h" />
     <ClInclude Include="DiscordIntegration.h" />
     <ClInclude Include="DisplayLayoutScreen.h" />
