+++ UI/UI.vcxproj.filters
@@ -89,6 +89,9 @@
     <ClCompile Include="RetroAchievementScreens.cpp">
       <Filter>Screens</Filter>
     </ClCompile>
+    <ClCompile Include="DebugOverlay.cpp">
+      <Filter>Screens</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GameInfoCache.h" />
@@ -176,6 +179,9 @@
       <Filter>Screens</Filter>
     </ClInclude>
     <ClInclude Include="RetroAchievementScreens.h">
+      <Filter>Screens</Filter>
+    </ClInclude>
+    <ClInclude Include="DebugOverlay.h">
       <Filter>Screens</Filter>
     </ClInclude>
   </ItemGroup>
