+++ UWP/UI_UWP/UI_UWP.vcxproj.filters
@@ -36,6 +36,7 @@
     <ClCompile Include="..\..\UI\CustomButtonMappingScreen.cpp" />
     <ClCompile Include="..\..\UI\TabbedDialogScreen.cpp" />
     <ClCompile Include="..\..\UI\RetroAchievementScreens.cpp" />
+    <ClCompile Include="..\..\UI\DebugOverlay.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h" />
@@ -73,5 +74,6 @@
     <ClInclude Include="..\..\UI\CustomButtonMappingScreen.h" />
     <ClInclude Include="..\..\UI\TabbedDialogScreen.h" />
     <ClInclude Include="..\..\UI\RetroAchievementScreens.h" />
+    <ClInclude Include="..\..\UI\DebugOverlay.h" />
   </ItemGroup>
 </Project>
